Fix naming for feedback package
